Some minor changes, plus temp files
This patch adds a couple of small content changes to workflow.tex,
a little bit of a document change to the design requirement template
and a consequent change in the template generating script, as well as
hierarchy.xml. The purpose of this document is to provide a
high-level overview of the design requirements. They are separated
into functional groups, and the ones which are currently tasked as
high priority are tagged with "base-functionality".