Add SerializerConfig module and test
This patch adds the SerializerConfig module. This module has a simple
API and is responsible for handling the implementation details of the
Serializer modules. To run this module's tests, run `make
SerializerConfigTest`. Any failing tests will appear in the output.
include/RuntimeLibrary/Serializer/SerializerConfig.h is documented
with the module API.