Add ScrumExcept, ScrumContext
ScrumExcept is the exception handling system for Autoscrum. It
functions very similarly to the stdc 'errno' and 'strerror' system.
ScrumContext is a wrapper for this data. Whenever a module
requires persistent data from the user, it should take a pointer to
one of these. When an error occurs, follow the procedure outlined in
ScrumExcept.h to report it.