Update SerializerConfig module to new conventions
This patch updates the SerializerConfig module to use the
conventions changed in the previous patch. All module functions now
take a pointer to ScrumExcept, which contains a struct
SerializerConfig. In addition, the module also correctly reports
errors to the user, and a test was added to verify that this is the
case.